example-facebook-post-weekly/requirements.txt
requests==2.2.1
simplejson==3.3.3
facebook-sdk==0.4.0

# eof
